pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.1",
    "openai>=1.76.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "together>=1.5.7",
]
